ai-service/requirements.txt
torch
torchvision
opencv-python
pillow
fastapi
uvicorn
python-multipart
numpy
requests
ultralytics